Updated all routes. Ref and Que model to be added.

+++ routes/chapter.js
@@ -1,9 +1,12 @@
 const express = require('express');
 const router = express.Router();
+
 const Part = require('../models/part');
 const Chapter = require('../models/chapter');
 const Section = require('../models/section');
 const { getIndex } = require('../utils/getIndex');
+const { getLinkObject } = require('../utils/getLinkObject');
+const { link } = require('joi');
 
 // get route to each chapter page
 router.get('/:id', async (req, res) => {
@@ -13,11 +16,17 @@ router.get('/:id', async (req, res) => {
     // find chapter document
     const chapter = await Chapter.findById(chapterID);
 
+    // get next and previous chapter if they exist
+    const nextChapter = await Chapter.findOne({ index: chapter.index + 1, part: chapter.part });
+    const previousChapter = await Chapter.findOne({ index: chapter.index - 1, part: chapter.part });
+
     // find all sections in chapter
     const sections = await Section.find({ chapter: chapterID });
 
+    const linkObject = await getLinkObject(chapter, 'chapter');
+
     // render chapter.ejs with chapter variable
-    res.render('chapter/show', { chapter, sections });
+    res.render('chapter/show', { chapter, sections, linkObject, nextChapter, previousChapter });
 });
 
 // get route to edit page of each chapter
@@ -28,8 +37,9 @@ router.get('/:id/edit', async (req, res) => {
     // find chapter document
     const chapter = await Chapter.findById(chapterID);
 
+    const linkObject = await getLinkObject(chapter, 'chapter');
     //render chapter_edit.ejs file with chapter variable
-    res.render('chapter/edit', { chapter });
+    res.render('chapter/edit', { chapter, linkObject });
 });
 
 // post route to edit chapter
@@ -54,17 +64,27 @@ router.post('/:id', async (req, res) => {
 router.delete('/:id', async (req, res) => {
     const chapterID = req.params.id;
 
-    // find chapter document
-    const chapter = await Chapter.findById(chapterID);
+    // find section documents
+    const sections = await Section.find({ chapter: chapterID });
 
-    // find part document
-    const part = await Part.findById(chapter.part);
+    // if sections exist, send error message
+    if (sections.length > 0) {
+        res.send('Cannot delete chapter with sections');
+    }
+    else {
 
-    // delete chapter document
-    await Chapter.findByIdAndDelete(chapterID);
+        // find chapter document
+        const chapter = await Chapter.findById(chapterID);
 
-    // redirect to part page
-    res.redirect(`/pola/subject/${part._id}`);
+        // find part document
+        const part = await Part.findById(chapter.part);
+
+        // delete chapter document
+        await Chapter.findByIdAndDelete(chapterID);
+
+        // redirect to part page
+        res.redirect(`/pola/subject/${part._id}`);
+    }
 });
 
 // post route to add section to chapter
