+++ routes/part.js
@@ -5,7 +5,7 @@ const Part = require('../models/part');
 const Subject = require('../models/subject');
 const Chapter = require('../models/chapter');
 const { getIndex } = require('../utils/getIndex');
-
+const { getLinkObject } = require('../utils/getLinkObject');
 
 // get route to each part page
 router.get('/:id', async (req, res) => {
@@ -17,8 +17,10 @@ router.get('/:id', async (req, res) => {
     // find all chapters in part
     const chapters = await Chapter.find({ part: partID });
 
+    const linkObject = await getLinkObject(part, 'part');
+
     // render part.ejs with part variable
-    res.render('part/show', { part, chapters });
+    res.render('part/show', { part, chapters, linkObject });
 });
 
 
@@ -36,8 +38,10 @@ router.get('/:id/edit', async (req, res) => {
     // get subject
     const subject = await Subject.findById(part.subject);
 
+    const linkObject = await getLinkObject(part, 'part');
+
     // render part_edit.ejs file with part and parts variables
-    res.render('part/edit', { part, parts, subject });
+    res.render('part/edit', { part, parts, subject, linkObject });
 });
 
 // post route to subject part quick-edit part
@@ -50,26 +54,46 @@ router.post('/:id', async (req, res) => {
     // update part document with new data
     part.title = req.body.title;
     part.summary = req.body.summary;
+    part.body = req.body.body;
 
     // save updated part document
     await part.save();
 
-    // redirect to subject page
-    res.redirect(`/pola/${part.subject}`);
+    // redirect to part page
+    res.redirect(`/pola/subject/${part._id}/edit`);
 });
 
 // delete route to delete part
 router.delete('/:id', async (req, res) => {
     const partID = req.params.id;
 
-    //find part document
-    const part = await Part.findById(partID);
+    // find chapter document related to part
+    const chapters = await Chapter.find({ part: partID });
 
-    // delete part document
-    await Part.findByIdAndDelete(partID);
+    // if chapters exist, then send error message
+    if (chapters.length > 0) {
+        res.send('Cannot delete part with chapters.');
+    }
+    else {
 
-    // redirect to subject page
-    res.redirect(`/pola/${part.subject}`);
+        //find part document
+        const part = await Part.findById(partID);
+
+        // get all parts with index greater than part index
+        const parts = await Part.find({ subject: part.subject, index: { $gt: part.index } });
+
+        // loop through parts and decrement index
+        for (let i = 0; i < parts.length; i++) {
+            parts[i].index--;
+            await parts[i].save();
+        }
+
+        // delete part document
+        await Part.findByIdAndDelete(partID);
+
+        // redirect to subject page
+        res.redirect(`/pola/${part.subject}`);
+    }
 });
 
 
@@ -91,8 +115,13 @@ router.post('/:id/chapter', async (req, res) => {
     // save new chapter document
     await newChapter.save();
 
-    // redirect to part page
-    res.redirect(`/pola/subject/${partID}`);
+    // redirect to part page or last chapter page if part has chapters
+    if (chapters.length > 0) {
+        res.redirect(`/pola/subject/part/${chapters[chapters.length - 1]._id}`);
+    }
+    else {
+        res.redirect(`/pola/subject/${partID}`);
+    }
 });
