+++ routes/subject.js
@@ -3,6 +3,7 @@ const router = express.Router();
 const Subject = require('../models/subject');
 const Part = require('../models/part');
 const { getIndex } = require('../utils/getIndex');
+const { getLinkObject } = require('../utils/getLinkObject');
 
 // get route on home page
 router.get('/', async (req, res) => {
@@ -11,7 +12,10 @@ router.get('/', async (req, res) => {
     const subjects = await Subject.find()
         .sort({ index: 1 });
 
-    res.render('subject/home', { subjects });
+    const linkObject = [];
+    // render subject_home.ejs file with subjects variable
+    res.render('subject/home', { subjects, linkObject });
+
 });
 
 
@@ -21,12 +25,17 @@ router.post('/', async (req, res) => {
     // get subjects
     const subjects = await Subject.find();
 
+    let imageURL = req.body.image;
+    if (req.body.image === '') {
+        imageURL = 'https://www.wolflair.com/wp-content/uploads/2017/01/placeholder.jpg';
+    };
+
     // create new content document
     const newSubject = new Subject({
         title: req.body.title,
         summary: req.body.summary,
         index: getIndex(subjects),
-        image: req.body.image
+        image: imageURL
     });
 
     // save new content document
@@ -47,8 +56,10 @@ router.get('/:id/edit', async (req, res) => {
     // get all subjects
     const subjects = await Subject.find();
 
+    const linkObject = await getLinkObject(subject, 'subject');
+
     // render subject_edit.ejs file with subject and parts variables
-    res.render('subject/edit', { subject, subjects });
+    res.render('subject/edit', { subject, subjects, linkObject });
 });
 
 
@@ -59,71 +70,56 @@ router.post('/:id', async (req, res) => {
     // find subject document
     const subject = await Subject.findById(subjectID);
 
+    // get subject with same index
+    const subjectToSwap = await Subject.find({ index: req.body.index });
+
+    // swap index of subject and subjectToSwap
+    subjectToSwap[0].index = subject.index;
+
     // update subject document with new data
     subject.title = req.body.title;
     subject.summary = req.body.summary;
     subject.image = req.body.image;
+    subject.index = req.body.index;
+    subject.body = req.body.body;
 
     // save updated subject document
     await subject.save();
+    // save updated subjectToSwap document
+    await subjectToSwap[0].save();
 
     // redirect to home page
-    res.redirect('/pola');
+    res.redirect(`/pola/${subjectID}`);
 });
 
 // delete route to delete subject
 router.delete('/:id', async (req, res) => {
     const subjectID = req.params.id;
 
-    // delete subject document
-    await Subject.findByIdAndDelete(subjectID);
+    // query to find all part documents with reference to subjectID
+    const parts = await Part.find({ subject: subjectID });
 
-    // redirect to home page
-    res.redirect('/pola');
-});
+    // if parts exist, send error message
+    if (parts.length > 0) {
+        res.send('Cannot delete. Subject has parts.');
+    }
+    else {
 
-// post route to reduce index of subject
-router.post('/:id/reduce', async (req, res) => {
-    const subjectID = req.params.id;
+        // find all subjects with index greater than subject index
+        const subjects = await Subject.find({ index: { $gt: req.body.index } });
 
-    // find subject document
-    const subject = await Subject.findById(subjectID);
+        // decrease index of all subjects with index greater than subject index
+        subjects.forEach(async (subject) => {
+            subject.index -= 1;
+            await subject.save();
+        });
 
-    // find subject document with index one less than subject
-    const subjectToSwap = await Subject.find({ index: subject.index - 1 });
+        // delete subject document
+        await Subject.findByIdAndDelete(subjectID);
 
-    // swap index of subject and subjectToSwap
-    subject.index = subject.index - 1;
-    subjectToSwap[0].index = subjectToSwap[0].index + 1;
-
-    // save updated subject documents
-    await subject.save();
-    await subjectToSwap[0].save();
-
-    // redirect to home page
-    res.redirect('/pola');
-});
-
-// post route to increase index of subject
-router.post('/:id/increase', async (req, res) => {
-    const subjectID = req.params.id;
-
-    // find subject document
-    const subject = await Subject.findById(subjectID);
-
-    // find subject document with index one less than subject
-    const subjectToSwap = await Subject.find({ index: subject.index + 1 });
-
-    // swap index of subject and subjectToSwap
-    subject.index = subject.index + 1;
-    subjectToSwap[0].index = subjectToSwap[0].index - 1;
-
-    // save updated subject documents
-    await subject.save();
-    await subjectToSwap[0].save();
-
-    // redirect to home page
-    res.redirect('/pola');
+        // redirect to home page
+        res.redirect('/pola');
+    }
 });
 
 // get route to each subject page
@@ -135,8 +131,10 @@ router.get('/:id', async (req, res) => {
     // find all part documents with reference to subjectID
     const parts = await Part.find({ subject: subjectID })
 
+    const linkObject = await getLinkObject(subject, 'subject');
+
     // render subject.ejs with subject and parts variables
-    res.render('subject/show', { subject, parts });
+    res.render('subject/show', { subject, parts, linkObject });
 });
 
 //post route to add new part
